+++ packages/ui/src/reviews/index.ts
@@ -1,0 +1,7 @@
+import { ReviewsSection } from './ReviewsSection'
+import type { IReviewsItem } from './ReviewsItem'
+
+export {
+  ReviewsSection,
+  IReviewsItem
+}
