+++ packages/ui/src/reviews/styles.css
@@ -23,11 +23,11 @@
   content: '' !important;
 }
 
-.reviewsSwiper .swiper-button-next svg,
+/* .reviewsSwiper .swiper-button-next svg,
 .reviewsSwiper .swiper-button-prev svg {
   width: 24px !important;
   height: 24px !important;
-}
+} */
 
 .reviewsSwiper > .swiper-wrapper {
   @apply !py-24 !px-8;
@@ -35,6 +35,6 @@
 
 @media (min-width: 1280px) {
   .swiper-wrapper:nth-child(1) {
-    padding-left: 40% !important;
+    padding-left: 20% !important;
   }
 }
